requirements.txt
PyQt6==6.6.1
pyinstaller==6.3.0
Pillow==10.2.0 